agent/requirements.txt
firebase-admin>=6.5.0
google-cloud-firestore>=2.11.1
python-dotenv>=1.0.0
psutil>=5.9.0
